modifico aplicacion de error ganancia y potencia

+++ common/src/lib.rs
@@ -37,7 +37,7 @@ pub struct EchosounderParameters {
     pub pulse_repetition_interval: f64, // Hz, en simulation lo convertimos a segundos.
     pub uses_high_frecuency: bool,
     pub transmited_potency: f64,
-    pub gain: f32,
+    pub gain: f64,
     pub threshold: f64,
     pub sound_speed: f64,
 }
